Initial commit to move processes inside docker containers (#2)
* Initial commit to move processes inside docker containers

The postgres successfully sets itself up. The django app isn't quite there yet. I am having docker build context issues because in the Dockerfile I am adding files from outside the Dockerfile directory. I could just clone the repo in and run it that way but I don't think that's the right model for this yet because I want to be able to make local changes and test them without pushing.

The django container still needs to be changed to look for the database on the docker network.

I also decided to go with docker-compose instead of ansible for two reasons

1) Eventually this will deploy with AWS Code deploy (finder crossed) so docker-compose/ansible won't matter
2) docker-compose can bring containers up/down much more gracefully that ansible. It also automatically creates a network both containers run on so they are linked without having to manually link them in ansible code. For doing docker-specific work docker-compose is a better tool

* Removing proxy stuff. That should come from deployment and not be baked in.

* Fixing docker-compose

To run:

set DATABASE variable equal to the hostname of the VM you are loading this in. Ensure ports 12345 and 8000 are open on your firewall, and then run docker-compose up --build